+++ composer.json
@@ -11,7 +11,7 @@
     ],
     "authors": [
         {
-            "name": "DNA Design",
+            "name": "DNA Designed Communications Limited",
             "email": "dev@dna.co.nz"
         }
     ],
